Fixed adding existing item to cart not increment quantity - Fixes #6
